Added icons for delete and update on frontend

+++ backend/controller/ItemController.js
@@ -226,15 +226,19 @@ export const updateItems = async (req, res) => {
         const file = req.files.file;
         const fileSize = file.data.length;
         const ext = path.extname(file.name);
-        fileName = file.md5 + ext;
+        const randomName = `${Date.now()}-${Math.round(Math.random() * 100000)}`
+        fileName = `${randomName}${ext}`;
+     //   fileName = file.md5 + ext;
         const allowedType = ['.png', '.jpg', '.jpeg'];
 
         if (!allowedType.includes(ext.toLowerCase())) return res.status(422).json({ msg: "Invalid Images" });
         if (fileSize > 5000000) return res.status(422).json({ msg: "Image must be less than 5 MB" });
 
         const filepath = `./public/images/${items.image}`;
-        fs.unlinkSync(filepath);
-
+        if (fs.existsSync(filepath)) {
+            fs.unlinkSync(filepath);
+        }
+    
         file.mv(`./public/images/${fileName}`, (err) => {
             if (err) return res.status(500).json({ msg: err.message });
         });
